Speed up the UserInvite integration test
Sped up the test by mock-asserting the send_msg_email is sent. The test
was slow because it was waiting for a response from the Email client to
know if the statement executed successfully or not. Ideally sending
emails should be done in the background and mocked automatically in the
tests (maybe someday). However, asserting the message is sent would need
to be done regardless.
